requirements.txt
tensorflow==2.19.0
keras==3.10.0
numpy==2.0.2
matplotlib==3.10.0
seaborn==0.13.2
gradio
tensorflow-hub==0.16.1
kagglehub==0.3.13
scikit-learn==1.6.1
tensorflowjs==4.22.0
Pillow==12.1.0
pandas==2.2.2
